In current implementation `attach_to_template()` runs one API call to attach each device to a template. However common deployment scenario is to attach the multiple devices to one device template and for those scenarios we can run one API call with payload for all the device-uuids. It improves the performance of SDK a lot, for example see the below time analysis. 

**Before fix:**

time vmanage import attachments -f ../python-viptela/devnet-sandbox-variables.yaml
Importing attachments from ../python-viptela/devnet-sandbox-variables.yaml
Attachment Updates: 11
vmanage import attachments -f ../python-viptela/devnet-sandbox-variables.yaml  1.26s user 0.09s system 0% cpu **3:39.68** total

**After fix:**

time vmanage import attachments -f devnet-sandbox-variables.yaml
Importing attachment(s) from devnet-sandbox-variables.yaml
Attachment Updates: 11
vmanage import attachments -f devnet-sandbox-variables.yaml  0.68s user 0.10s system 0% cpu **2:34.64** total

In the above example, I had total 11 devices in deployment with 8 devices attached to one device template and 3 devices attached to 3 device templates respectively.
 
In the current code we would execute **11** API calls because it is one API call per device. In the new code we would execute **4** API calls because it is executed per template with all devices attached to that template.  

PR contains the code needed for enhancement #76